divide todolist model logic
