+++ src/scripts/parts/todolist.js
@@ -1,5 +1,6 @@
 import $ from "jquery";
 import { createItem } from './todolist/view';
+import { getTodolist, storeNote, setFinishTodolist } from './todolist/model';
 
 function init() {
   console.log("Hello World! Todolist!");
@@ -16,7 +17,10 @@ function init() {
 function addNote() {
   var inputValue = $("#add-text").val();
   if (inputValue != "") {
-    storeNote(inputValue);
+    storeNote(inputValue, function(params){
+      var {id, text, status} = params;
+      $("ul").append(createItem(id, text, status));
+    });
     $("#add-text").val("");
   }
 }
@@ -75,8 +79,7 @@ function editItem() {
 function deleteItem() {
   var id = $(this).attr("value");
   $("li." + id).remove();
-  var noteArray = [];
-  noteArray = getTodolist();
+  var noteArray = getTodolist();
   for (var i = 0; i < noteArray.length; i++) {
       if (id == noteArray[i].id) {
           noteArray.splice(i, 1);
@@ -96,23 +99,8 @@ function checkItem(id) {
   }
 }
 
-//api return all todo list in an array
-function getTodolist() {
-  var Todolist = [];
-  Todolist = JSON.parse(localStorage.getItem("noteStr"));
-
-  return Todolist;
-}
-//api give an id and return event status
-function finishTodolist(id) {
-  var TodolistJSON;
-  TodolistJSON = JSON.parse(localStorage.getItem(id));
-  return TodolistJSON.status;
-}
-
 function initPage() {
-  var noteArray = [];
-  noteArray = getTodolist();
+  var noteArray = getTodolist();
   $("ul").empty();
   if (noteArray != null) {
     for (var i = 0; i < noteArray.length; i++) {
@@ -125,47 +113,4 @@ function initPage() {
   }
 }
 
-function storeNote(text) {
-  var noteArray = [];
-  var createTime = getCurrentTime();
-  var id = 1;
-  if (localStorage.getItem("noteStr") != null) {
-    noteArray = JSON.parse(localStorage.getItem("noteStr"));
-    id = noteArray[noteArray.length - 1].id + 1;
-  }
-  var noteJson = {
-    id: id,
-    name: text,
-    status: false,
-    createAt: createTime,
-    finishAt: null
-  };
-  noteArray.push(noteJson);
-  localStorage.setItem("noteStr", JSON.stringify(noteArray));
-  $("ul").append(createItem(id, text, false));
-}
-
-function setFinishTodolist(id) {
-  var noteArray = getTodolist();
-  for (var i = 0; i < noteArray.length; i++) {
-      if (noteArray[i].id == id) {
-          noteArray[i].status = true;
-          localStorage.setItem("noteStr", JSON.stringify(noteArray));
-          return true;
-      }
-  }
-  return false;
-}
-
-function getCurrentTime() {
-  var today = new Date();
-  var date =
-      today.getFullYear() + "-" + (today.getMonth() + 1) + "-" + today.getDate();
-  var time =
-      today.getHours() + ":" + today.getMinutes() + ":" + today.getSeconds();
-  var dateTime = date + " " + time;
-
-  return dateTime;
-}
-
-export { init, finishTodolist, setFinishTodolist };
+export { init, getTodolist, setFinishTodolist };
